@@ -1,4 +1,5 @@
-﻿using System.Security.Principal;
+﻿using System;
+using System.Security.Principal;
 
 namespace Enigmatry.Blueprint.Api.Tests.Infrastructure.TestImpersonation
 {
@@ -6,12 +7,14 @@ namespace Enigmatry.Blueprint.Api.Tests.Infrastructure.TestImpersonation
     {
         private const string TestUsername = "test_user@xyz.com";
 
-        private TestPrincipal(string userName)
+        private TestPrincipal(Guid userId, string userName)
         {
+            UserId = userId;
             UserName = userName;
             Identity = new GenericIdentity(userName);
         }
 
+        public Guid UserId { get; }
         public string UserName { get; }
 
         public bool IsInRole(string role)
@@ -24,7 +27,7 @@ namespace Enigmatry.Blueprint.Api.Tests.Infrastructure.TestImpersonation
 
         public static TestPrincipal CreateDefault()
         {
-            return new TestPrincipal(TestUsername);
+            return new TestPrincipal(Guid.NewGuid(), TestUsername);
         }
     }
 }
